pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^3.0"
python = "^3.8"
django-latex = "^0.1"
pypdf2 = "^3.0.1"
openai = "^0.27.8"
python-docx = "^0.8.11"
django-tex = "^1.1.10"
pdflatex = "^0.1.3"
pylatex = "^1.4.1"
[tool.poetry.dev-dependencies]
